export fetch-product to reuse across site
